Add the 'ATF1_idx' parameter
- Added a parameter for the position of the ATF1 protein.

- The parameter can either be an integer (the position on the polymer)
or 'None' (default), in which case no ATF1 is present.

- Added the function 'none_or_int' in module 'parameters.py' to handle
the different possible input types. In the case that 'None' is input, it
converts it to a NoneType object. Integer values are the only other
allowed values.

- Updated all modules to reflect these changes.

- Changed the 'cenH_indices' variable to 'const_silent_indices'. These
indices now include the ATF1 index (if relevant).

- Additional changes are needed: the functions for creating filenames
and plot titles have to be updated to include information on ATF1.